pyproject.toml
[tool.poetry]
name = "-storage"
version = "0.1.0"
description = ""
authors = ["Steve <Stigmata2m53@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
